pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["Andrii Khirilov <khirilov.and@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2"
python-dotenv = "^0.20.0"
django-rest-framework = "^0.1.0"
psycopg2 = "^2.9.3"
dj-rest-auth = "^2.2.5"
djangorestframework-simplejwt = "^5.2.0"
django-rest-auth = "^0.9.5"
djangorestframework-jwt = "^1.11.0"
django-rest-swagger = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
